+++ wxc_sdk/workspace_settings/__init__.py
@@ -4,6 +4,7 @@ Workspace settings, mostly identical to user settings
 """
 from dataclasses import dataclass
 
+from .devices import WorkspaceDevicesApi
 from .numbers import WorkspaceNumbersApi
 from ..api_child import ApiChild
 from ..person_settings.call_intercept import CallInterceptApi
@@ -27,22 +28,24 @@ class WorkspaceSettingsApi(ApiChild, base='workspaces'):
     this class are instances of the respective user settings APIs. When calling endpoints of these APIs workspace IDs
     need to be passed to the ``person_id`` parameter of the called function.
     """
-    call_intercept: CallInterceptApi
+    forwarding: PersonForwardingApi
     call_waiting: CallWaitingApi
     caller_id: CallerIdApi
-    forwarding: PersonForwardingApi
     monitoring: MonitoringApi
     numbers: WorkspaceNumbersApi
     permissions_in: IncomingPermissionsApi
     permissions_out: OutgoingPermissionsApi
+    devices: WorkspaceDevicesApi
+    call_intercept: CallInterceptApi
 
     def __init__(self, session: RestSession):
         super().__init__(session=session)
-        self.call_intercept = CallInterceptApi(session=session, workspaces=True)
+        self.forwarding = PersonForwardingApi(session=session, workspaces=True)
         self.call_waiting = CallWaitingApi(session=session, workspaces=True)
         self.caller_id = CallerIdApi(session=session, workspaces=True)
-        self.forwarding = PersonForwardingApi(session=session, workspaces=True)
         self.monitoring = MonitoringApi(session=session, workspaces=True)
         self.numbers = WorkspaceNumbersApi(session=session)
         self.permissions_in = IncomingPermissionsApi(session=session, workspaces=True)
         self.permissions_out = OutgoingPermissionsApi(session=session, workspaces=True)
+        self.devices = WorkspaceDevicesApi(session=session)
+        self.call_intercept = CallInterceptApi(session=session, workspaces=True)